File restructuring.
