+++ package.json
@@ -7,7 +7,7 @@
     "node": ">=6.9.1"
   },
   "scripts": {
-    "start": "babel-node ./app.js",
+    "start": "babel-node ./src/index.js",
     "test": "mocha"
   },
   "dependencies": {
