+++ src/lib/logger.js
@@ -1,0 +1,9 @@
+import bunyan from 'bunyan';
+
+const logger = bunyan.createLogger({
+    name: 'cotoami',
+    serializers: bunyan.stdSerializers
+});
+
+export default logger;
+
